Add benchmarks for FastReplace and GetCountry methods
Added a new benchmark method `FastReplace` in the
`StringExtensionsBenchmark` class to test the
`StringExtensions.FastReplace` method.

Added a new benchmark method `GetCountry` in the
`CountryRepositoryBenchmark` class to test the
`CountryRepository.GetCountry` method.

Updated metadata for `FastReplace` and `GetCountry` methods
to reflect changes in `UnitTestStatus` and `BenchmarkStatus`.

Updated the `Last Modified On` metadata in `CountryRepository.cs`
to `10-23-2025`.